file-management: Fixed error when id is invalid

+++ desofsApi/src/main/java/isep/ipp/pt/api/desofs/Service/ReceitaService/ReceitaServiceImpl.java
@@ -159,6 +159,9 @@ public class ReceitaServiceImpl implements ReceitaService{
 
     @Override
     public void deleteById(Long id) {
+        if(receitaRepo.findbyId(id) == null){
+            return;
+        }
         String filePathStr = receitaRepo.findbyId(id).getPath();
         deleteFile(filePathStr);
         receitaRepo.deleteById(id);
